Add files via upload
Create a bank entry form using appropriate form elements. The account
number must not be visible on the screen. The name and address
must be stored in one place. There must be a text box showing the
opening balance of the customer. The user should be able to make a
choice of either a deposit (or) withdrawal transaction. Accordingly,
when the user deposits (or) withdraws money, the opening balance
must be updated using CREDIT/DEBIT button. The user should not be
able to make any entries in the opening balance text box.