Add files via upload
Using functions, write a JavaScript code that accepts user name and
password from user. Check their correctness and display appropriate
alert messages. Restrict the user to try only for a maximum of three
times
Modify the above program that suits the following requirements:
a. After a lapse of 15 seconds, the password should be generated
automatically.

b. For impatient users, place a button that displays the passwords
immediately without waiting for 15 seconds