Add files via upload
Design an XML document to store information about a student in an
engineering college affiliated to ANU. The information must include
college id, Name of the College, Brach, Year of Joining, and e-mail id.
Make up sample data for 3 students. Create a CSS style sheet and
use it to display the document.